feat: Complete VPN build and release system with improved structure
- Add build-release.sh script for creating distributable releases
- Reorganize project structure:
  - Move auxiliary files (Dockerfile, docker-compose.hub.yml, deny.toml) to templates/
  - Move performance benchmarks from benches/ to tests/benchmarks/
  - Move Docker configs from docker/ to templates/
  - Remove unused db/ directory (database created in /opt/vpn/db during installation)
- Fix vpn-identity module compilation:
  - Add SQLite support for build-time sqlx macros
  - Create simplified version for release builds
  - Add migrations for database schema
- Update install.sh:
  - Support symlink installation for binaries
  - Add templates directory installation
  - Create /opt/vpn/db during installation
  - Improve Docker files handling
- Create comprehensive uninstall.sh:
  - Remove all VPN components from /opt/vpn
  - Handle Docker containers and images
  - Backup database before removal
  - Clean up all dependencies
- Add "Back to main menu" navigation in VPN CLI
- Update .gitignore to include release artifacts
- Generate RELEASE_INFO.md with build details in release directory

The release system now creates a complete distributable archive with:
- All compiled binaries (vpn, vpn-proxy, vpn-identity)
- Configuration templates and Docker files
- Installation and uninstallation scripts
- Comprehensive documentation

+++ crates/vpn-identity/src/session.rs
@@ -54,12 +54,12 @@ impl SessionManager {
         let value = serde_json::to_string(&session)?;
         let expiry_secs = expiration.num_seconds() as usize;
         
-        self.redis.set_ex(&key, value, expiry_secs).await?;
+        self.redis.set_ex(&key, value, expiry_secs as u64).await?;
         
         // Also store in a user's session set for easy lookup
         let user_sessions_key = format!("{}user:{}", self.key_prefix, user_id);
         self.redis.sadd(&user_sessions_key, &session_id).await?;
-        self.redis.expire(&user_sessions_key, expiry_secs).await?;
+        self.redis.expire(&user_sessions_key, expiry_secs as i64).await?;
         
         Ok(session_id)
     }
@@ -83,7 +83,7 @@ impl SessionManager {
                 let updated_json = serde_json::to_string(&session)?;
                 let ttl: isize = self.redis.ttl(&key).await?;
                 if ttl > 0 {
-                    self.redis.set_ex(&key, updated_json, ttl as usize).await?;
+                    self.redis.set_ex(&key, updated_json, ttl as u64).await?;
                 }
                 
                 Ok(Some(session))
@@ -104,7 +104,7 @@ impl SessionManager {
         let current_ttl: isize = self.redis.ttl(&key).await?;
         if current_ttl > 0 {
             let new_ttl = current_ttl + extension.num_seconds() as isize;
-            self.redis.expire(&key, new_ttl as usize).await?;
+            self.redis.expire(&key, new_ttl as i64).await?;
         }
         
         Ok(())
